+++ tests/Pest.php
@@ -1,0 +1,5 @@
+<?php
+
+use FabioSerembe\BladeSVGPro\Tests\TestCase;
+
+pest()->extend(TestCase::class)->in('Feature');